% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{r_to_z}
\alias{r_to_z}
\title{Convert Pearson's r to Fisher's z}
\usage{
r_to_z(r)
}
\arguments{
\item{r}{Pearson's correlation}
}
\description{
Converts Pearson's r to Fisher's z for use in meta-analysis.
}
\examples{
r_to_z( c(.22, -.9, NA) )
}
