% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{round2}
\alias{round2}
\title{Round while keeping trailing zeroes}
\usage{
round2(x, digits = 2)
}
\arguments{
\item{x}{Numeric value to round}

\item{digits}{Digits for rounding}
}
\description{
Rounds a numeric value and formats it as a string, keeping trailing zeroes.
}
\examples{
round2(0.03000, digits = 4)

# compare to base round, which drops trailing zeroes and returns a numeric
round(0.03000, digits = 4)
}
