% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{z_to_r}
\alias{z_to_r}
\title{Convert Fisher's z to Pearson's r}
\usage{
z_to_r(z)
}
\arguments{
\item{z}{Fisher's z}
}
\description{
Converts Fisher's z to Pearson's r for use in meta-analysis.
}
\examples{
# convert Fisher's z of 1.1 to Pearson's r
z_to_r(1.1)
}
