% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{z_to_r}
\alias{z_to_r}
\title{Convert Fisher's z to Pearson's r}
\usage{
z_to_r(z)
}
\arguments{
\item{z}{Fisher's z}
}
\description{
Converts Fisher's z to Pearson's r for use in meta-analysis.
}
\examples{
z_to_r( c(1.1, NA, -0.2) )
}
